// src/utils/types/features.ts
export type Features = {
  countries?: string[];
  cities?: string[];
  districts?: string[];
  attractions?: string[];
  accommodation?: string[];
  supermarkets?: string[];
  shops?: string[];
  consumables?: string[];
  cafes?: string[];
  bakeries?: string[];
  gelaterias?: string[];
  restaurants?: string[];
  bars?: string[];
  nostalgiaEffect?: string[];
};

// export type Features = {
//   countries?: string;
//   cities?: string;
//   airports?: string;
//   accommodation?: string;
//   districtsNeighborhoods?: string;
//   attractions?: string;
//   parks?: string;
//   food?: string;
//   drink?: string;
//   restaurants?: string;
//   bars?: string;
//   cafes?: string;
//   bakeriesPatisseries?: string;
//   chocolatier?s: string;
//   gelaterias?: string;
//   markets?: string;
//   supermarkets?: string;
//   departmentStores?: string;
//   clothingAccessoryStores?: string;
//   interiorDesignStores?: string;
//   technologyStores?: string;
//   audioStores?: string;
//   kitchenStores?: string;
//   stationeryStores?: string;
//   bookStores?: string;
//   japaneseStores?: string;
//   bicycleStores?: string;
//   outdoorStores?: string;
//   tennisStores?: string;
//   uniqueElements?: string;
//   nostalgiaEffect?: string;
// };